[chore] 채팅 - 파일명 전면 수정, 컴포넌트 이름 함께 수정
