+++ app/_components/chats/GroupChatRoom.tsx
@@ -1,15 +1,23 @@
 "use client";
 
-import { useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import { useSession } from "next-auth/react";
+import { useQueryClient } from "@tanstack/react-query";
 import { supabase } from "@/utils/supabase/client";
 
 import type { ChatProps } from "@/app/_types/realtime-chats";
 
-import { useQueryClient } from "@tanstack/react-query";
+import { useGetMessagesList } from "@/app/_hooks/useQueries/chats";
 import { sendMessage } from "@/app/_api/messages/privateChat-api";
 
 import { QUERY_KEY_MESSAGES_LIST } from "@/app/_api/queryKeys";
+
+import {
+  changeRecruitingState,
+  countParticipants,
+  deleteParticipant,
+  getRecruitingNumber,
+} from "@/app/_api/messages/groupChat-api";
 
 import { Input } from "@nextui-org/react";
 import {
@@ -20,27 +28,24 @@ import {
   ModalFooter,
   Button,
 } from "@nextui-org/react";
-import { useGetMessagesList } from "@/app/_hooks/useQueries/chats";
 
-type ChatPropsExceptActionId = Omit<ChatProps, "actionId">;
-
-const PrivateChat = ({
+const GroupChatRoom = ({
   isOpen,
   onOpenChange,
   roomId,
-}: ChatPropsExceptActionId) => {
+  actionId,
+}: ChatProps) => {
   const [message, setMessage] = useState("");
-
-  // console.log("roomId", roomId);
-
   const queryClient = useQueryClient();
 
   // 현재 로그인한 유저 uid
   const session = useSession();
   const loggedInUserUid = session.data?.user.user_uid || "";
 
+  // console.log("roomId", roomId);
+
   useEffect(() => {
-    const subscription = supabase
+    const messageSubscription = supabase
       .channel(`${roomId}`)
       .on(
         "postgres_changes",
@@ -52,48 +57,33 @@ const PrivateChat = ({
             queryKey: [QUERY_KEY_MESSAGES_LIST],
           });
         },
-
-        // 시도 1. messages를 상태로 관리하기
-        // (payload) => {
-        //   // Update the specific part of the query data
-        //   const newMessage = payload.new;
-        //   setMessages((prevMessages) => [...prevMessages, newMessage]); // messages 업데이트
-        // },
-
-        // 시도 2. setQueryData 리팩토링 시도1
-        // (payload) => {
-        //   // Update the specific part of the query data
-        //   const newMessage = payload.new;
-        //   queryClient.setQueryData(
-        //     ["messagesList", loggedInUserUid],
-        //     (prevData) => {
-        //       [...prevData, newMessage];
-        //     },
-        //   );
-        // },
-
-        // 시도 3. setQueryData 리팩토링 시도2
-        // (payload) => {
-        //   // Update the specific part of the query data
-        //   const newMessage = payload.new;
-        //   queryClient.setQueryData(
-        //     ["messagesList"],
-        //     (prevData: MessageType[]) => {
-        //       // Check if prevData is an array, if not, initialize it as an empty array
-        //       const newData = Array.isArray(prevData)
-        //         ? [...prevData, newMessage]
-        //         : [newMessage];
-        //       return newData;
-        //     },
-        //   );
-        // },
       )
       .subscribe();
 
+    // postgres_changes 를 써서 참가자 테이블에 내 uid가 insert되면 'in'
+    // 내 uid가 delete되면 'out' 표시 할 수 있지 않을까??
+
+    // Presence 채널 구독
+    // const chatRoom = supabase.channel(`${roomId}`);
+
+    // const presenceSubscription = chatRoom
+    //   .on("presence", { event: "sync" }, () => {
+    //     const newState = supabase.channel(`${roomId}`).presenceState();
+    //     console.log("sync", newState);
+    //   })
+    //   .on("presence", { event: "join" }, ({ key, newPresences }) => {
+    //     console.log("join", key, newPresences);
+    //   })
+    //   .on("presence", { event: "leave" }, ({ key, leftPresences }) => {
+    //     console.log("leave", key, leftPresences);
+    //   })
+    //   .subscribe();
+
     return () => {
-      subscription.unsubscribe();
+      messageSubscription.unsubscribe();
+      // presenceSubscription.unsubscribe();
     };
-  }, [roomId]);
+  }, []);
 
   const { messagesList, isLoading, isError } = useGetMessagesList({
     roomId,
@@ -119,6 +109,29 @@ const PrivateChat = ({
     });
   };
 
+  // action 참여 취소 핸들러
+  const handleCancelParticipate = async (onClose: () => void) => {
+    const isConfirm = window.confirm("참여를 취소하시겠습니까?");
+    if (isConfirm) {
+      // 1. 채팅방 인원 === 모집인원 인지 확인하기
+      // (맞으면 내가 나갔을때 '모집중'으로 바꿔야 함)
+
+      // 현재 채팅방 인원 가져오기
+      const participantsNumber = await countParticipants(roomId);
+
+      // action 모집인원 가져오기
+      const recruitingNumber = await getRecruitingNumber(roomId);
+
+      if (participantsNumber === recruitingNumber) {
+        await changeRecruitingState({ action_id: actionId, mode: "out" });
+      }
+
+      // 2. 참가자 테이블에서 삭제
+      await deleteParticipant(loggedInUserUid);
+    }
+    onClose();
+  };
+
   return (
     <>
       {/* <Button onPress={onOpen}>Open Modal</Button> */}
@@ -131,8 +144,14 @@ const PrivateChat = ({
         <ModalContent className="max-w-[30%] h-[80%] overflow-y-auto scrollbar-hide">
           {(onClose) => (
             <>
-              <ModalHeader className="flex flex-col gap-1">
-                1:1 문의하기
+              <ModalHeader className="flex gap-1">
+                <span className="mr-5">action 참여자 단체 채팅방</span>
+                <button
+                  className="bg-black text-white px-2"
+                  onClick={() => handleCancelParticipate(onClose)}
+                >
+                  참여 취소하기
+                </button>
               </ModalHeader>
               <ModalBody>
                 <div className="flex justify-center">
@@ -184,4 +203,4 @@ const PrivateChat = ({
   );
 };
 
-export default PrivateChat;
+export default GroupChatRoom;
